% Generated by build script
\name{obis_checklist}
\alias{obis_checklist}
\title{Fetch OBIS species checklist}
\usage{
obis_checklist()
}
\arguments{
\item{scientificname}{Scientific name filter}
\item{taxonid}{Taxon ID filter}
\item{geometry}{WKT geometry filter}
\item{size}{Number of results (default 100)}
}
\value{tibble: taxonID, scientificName, taxonRank, family, records}
\description{Fetch OBIS species checklist}
